backend/requirements/base.txt
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
boto3==1.34.69
